requirements.txt
ib_insync
pandas
pandas-ta
numpy
python-dotenv
openai
vectorbt
pyyaml
finnhub-python
requests
fastapi
uvicorn[standard]
pytest
pytest-order
pytest-cov
httpx
psycopg2-binary

